Added general functions header
Added a general functions header to handle conversion between the time stored in seconds and the display the time in minutes/seconds on the LCD. These functions have also been implemented in main.c.

As a side note, I also added a set function for the on_time variable to allow for a global variable to store the user set on time between runs of the software (not saved between power cycles though).